requirements.txt
line-bot-sdk
requests
boto3